Hid bot token, made admin an array.
I put the gemfile in alphabetical order because I like it that way.

I added libsodium which you'll have to add the dll to the directory on your own since it's in the .gitignore

I made the bot grab the token from a yaml file which is also in the gitignore. In order to use the bot you'll have to create a file called config.yaml containing: token: 'bot token here'

Lastly, I changed it so admin is now an array of user id's so multiple users can use elevated commands.